Moved database to model
